refactor: pivot chart layout
