+++ packages/vtable/package.json
@@ -57,7 +57,7 @@
     "core-js": "3.31.1"
   },
   "devDependencies": {
-    "@visactor/vchart": "1.4.1-alpha.1",
+    "@visactor/vchart": "1.4.1",
     "@internal/bundler": "workspace:*",
     "@internal/eslint-config": "workspace:*",
     "@internal/ts-config": "workspace:*",
