+++ packages/vtable/src/event/pivot-chart/axis-hover.ts
@@ -1,5 +1,5 @@
 import { TABLE_EVENT_TYPE } from '../../core/TABLE_EVENT_TYPE';
-import type { PivotLayoutMap } from '../../layout/pivot-layout';
+import type { PivotHeaderLayoutMap } from '../../layout/pivot-header-layout';
 import type { BaseTableAPI } from '../../ts-types/base-table';
 
 export function bindAxisHoverEvent(table: BaseTableAPI) {
@@ -11,7 +11,7 @@ export function bindAxisHoverEvent(table: BaseTableAPI) {
     const cellGroup = table.scenegraph.getCell(e.col, e.row);
     cellGroup.forEachChildren(child => {
       if (child.name === 'axis') {
-        const position = getAxisPosition(e.col, e.row, table.internalProps.layoutMap as PivotLayoutMap);
+        const position = getAxisPosition(e.col, e.row, table.internalProps.layoutMap as PivotHeaderLayoutMap);
         (e as any).axisPosition = position;
         table.fireListeners(TABLE_EVENT_TYPE.MOUSEENTER_AXIS, e as any);
       }
@@ -22,7 +22,7 @@ export function bindAxisHoverEvent(table: BaseTableAPI) {
     const cellGroup = table.scenegraph.getCell(e.col, e.row);
     cellGroup.forEachChildren(child => {
       if (child.name === 'axis') {
-        const position = getAxisPosition(e.col, e.row, table.internalProps.layoutMap as PivotLayoutMap);
+        const position = getAxisPosition(e.col, e.row, table.internalProps.layoutMap as PivotHeaderLayoutMap);
         (e as any).axisPosition = position;
         table.fireListeners(TABLE_EVENT_TYPE.MOUSELEAVE_AXIS, e as any);
       }
@@ -30,7 +30,7 @@ export function bindAxisHoverEvent(table: BaseTableAPI) {
   });
 }
 
-function getAxisPosition(col: number, row: number, layout: PivotLayoutMap) {
+function getAxisPosition(col: number, row: number, layout: PivotHeaderLayoutMap) {
   if (layout.indicatorsAsCol) {
     if (
       layout.hasTwoIndicatorAxes &&
