+++ packages/vtable/src/layout/chart-helper/get-axis-config.ts
@@ -1,10 +1,10 @@
 import { isArray, isNumber, merge } from '@visactor/vutils';
-import type { PivotLayoutMap } from '../pivot-layout';
+import type { PivotHeaderLayoutMap } from '../pivot-header-layout';
 import type { ITableAxisOption } from '../../ts-types/component/axis';
 import type { PivotChart } from '../../PivotChart';
 import { getAxisDomainRangeAndLabels } from './get-axis-domain';
 
-export function getAxisConfigInPivotChart(col: number, row: number, layout: PivotLayoutMap): any {
+export function getAxisConfigInPivotChart(col: number, row: number, layout: PivotHeaderLayoutMap): any {
   if (!layout._table.isPivotChart()) {
     return undefined;
   }
@@ -32,14 +32,9 @@ export function getAxisConfigInPivotChart(col: number, row: number, layout: Pivo
       const data = layout.dataset.collectedValues[defaultKey + (isZeroAlign ? '_align' : '')]
         ? layout.dataset.collectedValues[defaultKey + (isZeroAlign ? '_align' : '')]
         : layout.dataset.collectedValues[defaultKey];
-      const index = layout.getRecordIndexByCol(col);
+      const colPath = layout.getColKeysPath(col);
 
-      const range = merge(
-        {},
-        (data?.[
-          layout.getColKeysPath()?.[index]?.[Math.max(0, layout.columnHeaderLevelCount - 1 - layout.topAxesCount)] ?? ''
-        ] as { min: number; max: number }) ?? { min: 0, max: 1 }
-      );
+      const range = merge({}, (data?.[colPath ?? ''] as { min: number; max: number }) ?? { min: 0, max: 1 });
       if (range.positiveMax && range.positiveMax > range.max) {
         range.max = range.positiveMax;
       }
@@ -103,13 +98,8 @@ export function getAxisConfigInPivotChart(col: number, row: number, layout: Pivo
       const data = layout.dataset.collectedValues[defaultKey + (isZeroAlign ? '_align' : '')]
         ? layout.dataset.collectedValues[defaultKey + (isZeroAlign ? '_align' : '')]
         : layout.dataset.collectedValues[defaultKey];
-      const index = layout.getRecordIndexByCol(col);
-      const range = merge(
-        {},
-        (data?.[
-          layout.getColKeysPath()?.[index]?.[Math.max(0, layout.columnHeaderLevelCount - 1 - layout.topAxesCount)] ?? ''
-        ] as { min: number; max: number }) ?? { min: 0, max: 1 }
-      );
+      const colPath = layout.getColKeysPath(col);
+      const range = merge({}, (data?.[colPath ?? ''] as { min: number; max: number }) ?? { min: 0, max: 1 });
       if (range.positiveMax && range.positiveMax > range.max) {
         range.max = range.positiveMax;
       }
@@ -180,9 +170,8 @@ export function getAxisConfigInPivotChart(col: number, row: number, layout: Pivo
       }
       const data = layout.dataset.collectedValues[rowDimensionKey] ?? ([] as string[]);
 
-      const recordRow = layout.getRecordIndexByRow(row);
-      const rowPath = layout.getRowKeysPath()[recordRow];
-      const domain = (data[rowPath?.[rowPath?.length - 1] ?? ''] as Array<string>) ?? [];
+      const rowPath = layout.getRowKeysPath(row);
+      const domain = (data[rowPath ?? ''] as Array<string>) ?? [];
 
       const { axisOption, isPercent } = getAxisOption(col + 1, row, 'left', layout);
       if (axisOption?.visible === false) {
@@ -221,13 +210,8 @@ export function getAxisConfigInPivotChart(col: number, row: number, layout: Pivo
       const data = layout.dataset.collectedValues[defaultKey + (isZeroAlign ? '_align' : '')]
         ? layout.dataset.collectedValues[defaultKey + (isZeroAlign ? '_align' : '')]
         : layout.dataset.collectedValues[defaultKey];
-      const index = layout.getRecordIndexByRow(row);
-      const range = merge(
-        {},
-        (data?.[
-          layout.getRowKeysPath()[index]?.[Math.max(0, layout.rowHeaderLevelCount - 1 - layout.leftAxesCount)] ?? ''
-        ] as { min: number; max: number }) ?? { min: 0, max: 1 }
-      );
+      const rowPath = layout.getRowKeysPath(row);
+      const range = merge({}, (data?.[rowPath ?? ''] as { min: number; max: number }) ?? { min: 0, max: 1 });
       if (range.positiveMax && range.positiveMax > range.max) {
         range.max = range.positiveMax;
       }
@@ -293,7 +277,7 @@ export function getAxisConfigInPivotChart(col: number, row: number, layout: Pivo
       row < layout.rowCount - layout.bottomFrozenRowCount
     ) {
       const indicatorKeys = layout.getIndicatorKeyInChartSpec(layout.rowHeaderLevelCount - 1, row);
-      let defaultKey = indicatorKeys[1];
+      let defaultKey = indicatorKeys?.[1];
       if (isArray(defaultKey)) {
         defaultKey = defaultKey[0];
       }
@@ -308,13 +292,8 @@ export function getAxisConfigInPivotChart(col: number, row: number, layout: Pivo
       const data = layout.dataset.collectedValues[defaultKey + (isZeroAlign ? '_align' : '')]
         ? layout.dataset.collectedValues[defaultKey + (isZeroAlign ? '_align' : '')]
         : layout.dataset.collectedValues[defaultKey];
-      const index = layout.getRecordIndexByRow(row);
-      const range = merge(
-        {},
-        (data?.[
-          layout.getRowKeysPath()[index]?.[Math.max(0, layout.rowHeaderLevelCount - 1 - layout.leftAxesCount)] ?? ''
-        ] as { min: number; max: number }) ?? { min: 0, max: 1 }
-      );
+      const rowPath = layout.getRowKeysPath(row);
+      const range = merge({}, (data?.[rowPath ?? ''] as { min: number; max: number }) ?? { min: 0, max: 1 });
       if (range.positiveMax && range.positiveMax > range.max) {
         range.max = range.positiveMax;
       }
@@ -377,9 +356,8 @@ export function getAxisConfigInPivotChart(col: number, row: number, layout: Pivo
       }
       const data = layout.dataset.collectedValues[columnDimensionKey] ?? ([] as string[]);
 
-      const recordCol = layout.getRecordIndexByCol(col);
-      const colPath = layout.getColKeysPath()[recordCol];
-      const domain = (data?.[colPath?.[colPath?.length - 1] ?? ''] as Array<string>) ?? [];
+      const colPath = layout.getColKeysPath(col);
+      const domain = (data?.[colPath ?? ''] as Array<string>) ?? [];
 
       const { axisOption, isPercent } = getAxisOption(col, row - 1, 'bottom', layout);
       if (axisOption?.visible === false) {
@@ -402,7 +380,7 @@ export function getAxisConfigInPivotChart(col: number, row: number, layout: Pivo
   return undefined;
 }
 
-export function getAxisOption(col: number, row: number, orient: string, layout: PivotLayoutMap) {
+export function getAxisOption(col: number, row: number, orient: string, layout: PivotHeaderLayoutMap) {
   const spec = layout.getRawChartSpec(col, row);
   if (spec && isArray(spec.axes)) {
     const axisOption = spec.axes.find((axis: any) => {
@@ -426,7 +404,7 @@ export function getAxisOption(col: number, row: number, orient: string, layout: 
   };
 }
 
-export function checkZeroAlign(spec: any, orient: string, layout: PivotLayoutMap) {
+export function checkZeroAlign(spec: any, orient: string, layout: PivotHeaderLayoutMap) {
   // check condition:
   // 1. two axes and one set sync
   // 2. axisId in sync is another
