+++ packages/vtable/src/scenegraph/group-creater/cell-type/chart-cell.ts
@@ -77,7 +77,7 @@ export function createChartCellGroup(
     //   y1: Math.ceil(cellGroup.globalAABBBounds.y1 + padding[0] + table.scrollTop),
     //   y2: Math.ceil(cellGroup.globalAABBBounds.y1 + height - padding[2] + table.scrollTop)
     // },
-    axes: table.internalProps.layoutMap.getChartAxes(col, row)
+    axes: table.isPivotChart() ? table.internalProps.layoutMap.getChartAxes(col, row) : []
     // clipRect: {
     //   left: cellGroup.globalAABBBounds.x1 + (table as any).tableX + padding[3],
     //   top: cellGroup.globalAABBBounds.y1 + (table as any).tableY + padding[0],
