+++ packages/vtable/src/tools/get-data-path/index.ts
@@ -1,8 +1,8 @@
 import { isArray } from '@visactor/vutils';
 import type { PivotChart } from '../../PivotChart';
-import { PivotLayoutMap } from '../../layout/pivot-layout';
 import type { IPivotTableCellHeaderPaths, PivotChartConstructorOptions } from '../../ts-types';
 import { createDataset } from './create-dataset';
+import { PivotHeaderLayoutMap } from '../../layout/pivot-header-layout';
 
 export function getDataCellPath(
   options: PivotChartConstructorOptions,
@@ -28,7 +28,7 @@ export function getDataCellPath(
   };
 
   // mock layoutMap
-  const layoutMap = new PivotLayoutMap(mockTable as PivotChart, dataset);
+  const layoutMap = new PivotHeaderLayoutMap(mockTable as PivotChart, dataset);
 
   // compare data
   for (let col = 0; col < layoutMap.colCount; col++) {
@@ -41,7 +41,7 @@ export function getDataCellPath(
       const aggregator = dataset.getAggregator(
         rowKey[rowKey.length - 1],
         colKey[colKey.length - 1],
-        (layoutMap as PivotLayoutMap).getIndicatorKey(col, row)
+        (layoutMap as PivotHeaderLayoutMap).getIndicatorKey(col, row)
       );
       const result = compareData(
         aggregator.value ? aggregator.value() : undefined,
@@ -64,7 +64,7 @@ function compareData(
   data2: Object,
   col: number,
   row: number,
-  layoutMap: PivotLayoutMap,
+  layoutMap: PivotHeaderLayoutMap,
   compareFunc?: (a: any, b: any) => boolean
 ) {
   if (isArray(data1)) {
